automatically backed up by learn

+++ db/migrate/001_create_costumes.rb
@@ -9,12 +9,12 @@
 # value types according to the spec
 
 class CreateCostumes < ActiveRecord::Migration
-  def change 
-    create_table :costumes do |t| 
-      t.string :name 
-      t.integer :price 
-      t.integer :size 
-      t.string :image_url 
-    end 
+  def change
+    create_table :costumes do |t|
+      t.string :name
+      t.float :price 
+      t.integer :size
+      t.string :image_url
+    end
   end
-end 
+end
